VIITE-3055 nodes and junctions interface
Api-rajapinta kaikkien voimassa oleville solmuille ja niihin liittyville liittymille

Korjattu kommenttien perusteella, erillinen tiedosto APIServicelle

Testattu
