+++ viite-backend/viite-main/src/main/scala/fi/liikennevirasto/viite/dao/LinearLocationDAO.scala
@@ -82,11 +82,6 @@ trait BaseLinearLocation {
   def endCalibrationPointType: CalibrationPointType =
     if (endCalibrationPoint.typeCode.isDefined) endCalibrationPoint.typeCode.get else CalibrationPointType.NoCP
 }
-case class JunctionCoordinate(xCoord: Double, yCoord: Double)
-
-case class JunctionWithLinearLocation(id: Long, junctionNumber: Option[Long], nodeNumber: Option[Long], startDate: DateTime, endDate: Option[DateTime],
-                                      validFrom: DateTime, validTo: Option[DateTime], createdBy: String, createdTime: Option[DateTime], llId: Seq[Long])
-
 
 case class CalibrationPointReference(addrM: Option[Long], typeCode: Option[CalibrationPointType] = Option.empty) {
   def isEmpty: Boolean = addrM.isEmpty
@@ -131,6 +126,14 @@ case class LinearLocation(id: Long, orderNumber: Double, linkId: String, startMV
   }
 
 }
+
+/**
+ * Coordinate point of junction (the blue circle on the map).
+ * Used only in /nodes/valid API
+ * @param xCoord
+ * @param yCoord
+ */
+case class JunctionCoordinate(xCoord: Double, yCoord: Double)
 
 //TODO Rename all the method names to follow a rule like fetchById instead of have fetchById and QueryById
 class LinearLocationDAO extends BaseDAO {
@@ -225,59 +228,6 @@ class LinearLocationDAO extends BaseDAO {
       val yCoord = r.nextDouble()
 
       JunctionCoordinate(xCoord, yCoord)
-    }
-  }
-
-  implicit val getJunctionWithLinearLocation: GetResult[JunctionWithLinearLocation] = new GetResult[JunctionWithLinearLocation] {
-    def parseJsonbAgg(stringToParse: String): Seq[Long] = {
-      val res = stringToParse.replaceAll("[\\[\\]\\s]", "")
-      res.split(",").map(_.toLong).toSeq
-    }
-
-    def parseBeforeAfterValue(beforeAfterToParse: String): String = {
-      // remove square brackets and spaces from the string i.e "[1, 2]" -> "1,2"
-      val beforeAfterParsed = beforeAfterToParse.replaceAll("[\\[\\]\\s]", "")
-      // Change beforeAfter value 1 to E "Ennen"
-      val beforeAfter1toE = beforeAfterParsed.replaceAll("1", "E")
-      // Change beforeAfter value 2 to J "Jälkeen"
-      val beforeAfter2toJ = beforeAfter1toE.replaceAll("2", "J")
-      beforeAfter2toJ
-    }
-
-    def apply(r: PositionedResult): JunctionWithLinearLocation = {
-      val id = r.nextLong()
-      val junctionNumber = r.nextLongOption()
-      val nodeNumber = r.nextLongOption()
-      val startDate = dateOptTimeFormatter.parseDateTime(r.nextDate.toString)
-      val endDate = r.nextDateOption.map(d => dateOptTimeFormatter.parseDateTime(d.toString))
-      val validFrom = dateOptTimeFormatter.parseDateTime(r.nextDate.toString)
-      val validTo = r.nextDateOption.map(d => dateOptTimeFormatter.parseDateTime(d.toString))
-      val createdBy = r.nextString()
-      val createdTime = r.nextDateOption.map(d => dateOptTimeFormatter.parseDateTime(d.toString))
-      val llIds: Seq[Long] = parseJsonbAgg(r.rs.getString("ll_id"))
-
-      JunctionWithLinearLocation(id, junctionNumber, nodeNumber, startDate, endDate,
-        validFrom, validTo, createdBy, createdTime, llIds)
-    }
-  }
-
-  def fetchJunctionsByNodeNumbersWithLinearLocation(nodeNumbers: Seq[Long]): Seq[JunctionWithLinearLocation] = {
-    if (nodeNumbers.isEmpty) {
-      Seq()
-    } else {
-      val query =
-        s"""
-    select distinct J.ID, J.JUNCTION_NUMBER, J.NODE_NUMBER, J.START_DATE, J.END_DATE, J.VALID_FROM, J.VALID_TO, J.CREATED_BY, J.CREATED_TIME, jsonb_agg(distinct ll.id) as ll_id
-              FROM JUNCTION J
-              join junction_point jp on j.id = jp.junction_id and jp.valid_to is null
-              join calibration_point cp on jp.roadway_point_id = cp.roadway_point_id and cp.valid_to is null
-              join linear_location ll on cp.link_id = ll.link_id and ll.valid_to is null
-              join roadway_point rp on cp.roadway_point_id = rp.id and rp.roadway_number = ll.roadway_number
-              join roadway r on ll.roadway_number = r.roadway_number and r.valid_to is null and r.end_date is null
-      where j.NODE_NUMBER in (${nodeNumbers.mkString(", ")}) AND j.VALID_TO IS NULL AND j.END_DATE IS NULL
-      group by j.id
-    """
-      Q.queryNA[JunctionWithLinearLocation](query).iterator.toSeq
     }
   }
 
@@ -464,8 +414,7 @@ class LinearLocationDAO extends BaseDAO {
     var intersectionFunction = s""
     var closingBracket = s""
 
-
-    if (llLength == 2) { //Tarkistetaan onko geometrian alku ja loppupisteet samat
+    if (llLength == 2) { //Check whether the geometry's start and endpoints are the same with 2 linear locations.
       val ll1 = allLL.find(_.id == llIds(0))
       val ll2 = allLL.find(_.id == llIds(1))
 
@@ -473,27 +422,19 @@ class LinearLocationDAO extends BaseDAO {
         val cr1 = crossingRoads.find(_.roadwayNumber == ll1.get.roadwayNumber)
         val cr2 = crossingRoads.find(_.roadwayNumber == ll2.get.roadwayNumber)
 
-        val sideCodeBeforeAfter1: (Int, Long) = (ll1.get.sideCode.value, cr1.get.beforeAfter)
-        val sideCodeBeforeAfter2: (Int, Long) = (ll2.get.sideCode.value, cr2.get.beforeAfter)
+        val beforeAfter1: Long = cr1.get.beforeAfter
+        val beforeAfter2: Long = cr2.get.beforeAfter
 
         var point = Point(0,0)
 
-        if (sideCodeBeforeAfter1._1 == 2 || sideCodeBeforeAfter2._1 == 2) {
-          // Atleast one sidecode is 2
-          point = if (sideCodeBeforeAfter1._1 == 2) {
-            if (sideCodeBeforeAfter1._2 == 1) ll1.get.getLastPoint
-            else ll1.get.getFirstPoint
-          } else {
-            if (sideCodeBeforeAfter2._2 == 1) ll2.get.getLastPoint
+        point =
+          (if (beforeAfter1 == beforeAfter2) { // If both have same before-after
+            if (beforeAfter1 == 1) ll1.get.getLastPoint
             else ll2.get.getFirstPoint
-          }
-        } else if (sideCodeBeforeAfter1._1 == 3 && sideCodeBeforeAfter2._1 == 3) {
-          // Both side codes are 3
-          point = if (sideCodeBeforeAfter1._2 == 1) ll1.get.getLastPoint
-          else if (sideCodeBeforeAfter1._2 == 2) ll1.get.getFirstPoint
-          else if (sideCodeBeforeAfter2._2 == 1) ll2.get.getLastPoint
-          else ll2.get.getFirstPoint
-        }
+          } else {
+            if (beforeAfter1 == 1) ll1.get.getLastPoint
+            else ll2.get.getLastPoint
+          })
         return Some(point)
       }
     }
